Move API to interfaces
