+++ pydeconz/api.py
@@ -1,64 +1,2 @@
-"""API base classes."""
-
-from __future__ import annotations
-
-from collections.abc import Awaitable, Callable, ItemsView, KeysView, ValuesView
-import logging
-from typing import Any
-
 from .models.api import APIItem  # noqa: F401
-
-LOGGER = logging.getLogger(__name__)
-
-
-class APIItems:
-    """Base class for a map of API Items."""
-
-    def __init__(
-        self,
-        raw: dict,
-        request: Callable[..., Awaitable[dict[str, Any]]],
-        path: str,
-        item_cls: Any,
-    ) -> None:
-        """Initialize API items."""
-        self._request = request
-        self._path = path
-        self._item_cls = item_cls
-        self._items: dict = {}
-        self.process_raw(raw)
-
-    async def update(self) -> None:
-        """Refresh data."""
-        raw = await self._request("get", self._path)
-        self.process_raw(raw)
-
-    def process_raw(self, raw: dict[str, Any]) -> None:
-        """Process data."""
-        for id, raw_item in raw.items():
-
-            if (obj := self._items.get(id)) is not None:
-                obj.update(raw_item)
-
-            else:
-                self._items[id] = self._item_cls(id, raw_item, self._request)
-
-    def items(self) -> ItemsView[str, Any]:
-        """Return items."""
-        return self._items.items()
-
-    def keys(self) -> KeysView[Any]:
-        """Return item keys."""
-        return self._items.keys()
-
-    def values(self) -> ValuesView[Any]:
-        """Return item values."""
-        return self._items.values()
-
-    def __getitem__(self, obj_id: str) -> Any:
-        """Get item value based on key."""
-        return self._items[obj_id]
-
-    def __iter__(self) -> Any:
-        """Allow iterate over items."""
-        return iter(self._items)
+from .interfaces.api import APIItems  # noqa: F401
